Qiskit.jl

Documentation for Qiskit.jl.

The documentation of Qiskit's C API may also be useful to reference when using this library.

API Reference

Qiskit.QuantumCircuit

Qiskit.SparseObservable

Qiskit.SparseObservable

Qiskit.Target

Qiskit.Target

Qiskit.TargetEntry

Qiskit.TranspileLayout

Qiskit.transpile

Qiskit.QuantumCircuit — Type

QuantumCircuit

Quantum circuit representation.

Available read-only properties:

num_qubits

num_clbits

num_instructions

data - contains instruction list

The additional properties are methods:

reset(qubit)

measure(qubit, clbit)

barrier(qubit1, qubit2, ...)

unitary(matrix, [qubit1, qubit2, ...])

many standard gates corresponding to Qiskit's Python API

Qiskit.SparseObservable — Type

SparseObservable

Qiskit observable. This is a wrapper of QkObs, which is similar to SparseObservable in Python.

Qiskit.SparseObservable — Method

SparseObservable(n::Integer)

Construct an empty SparseObservable on n qubits.

Qiskit.Target — Type

Target

A mapping of instructions and properties representing the particular constraints of a backend. Its purpose is to provide the compiler with information that allows it to compile an input circuit into another that is optimized taking in consideration the Target's specifications.

Available properties:

num_qubits

num_instructions

Qiskit.Target — Method

 Target(num_qubits)

Qiskit.TargetEntry — Type

TargetEntry

A mapping of qubit arguments and properties representing gate map of the Target.

Available properties:

num_properties

Qiskit.TranspileLayout — Type

TranspileLayout

This type stores the permutation introduced by the transpiler. In general Qiskit’s transpiler is unitary-preserving up to the initial layout and output permutations. The initial layout is the mapping from virtual circuit qubits to physical qubits on the target and the output permutation is caused by swap gate insertion or permutation elision prior to the initial layout being set in the transpiler pipeline. This type tracks these details and provide an interface to reason about these permutations.

Qiskit.transpile — Method

transpile(circuit, target)

Transpile a single circuit.

The Qiskit transpiler is a quantum circuit compiler that rewrites a given input circuit to match the constraints of a QPU and optimizes the circuit for execution.

This function wraps qk_transpile, which is multithreaded internally and will launch a thread pool with threads equal to the number of CPUs reported by the operating system by default. This will include logical cores on CPUs with simultaneous multithreading. You can tune the number of threads with the RAYON_NUM_THREADS environment variable. For example, setting RAYON_NUM_THREADS=4 would limit the thread pool to 4 threads.